Moved monad transformer code to new module
