+++ dunai-frp-bearriver/bearriver.cabal
@@ -35,6 +35,7 @@ description:
 library
   exposed-modules:     FRP.Yampa,
                        FRP.BearRiver
+                       FRP.BearRiver.Monad
 
   build-depends:       base >=4.6 && <5, transformers >=0.3, mtl, dunai >= 0.6.0 && < 0.8, MonadRandom, simple-affine-space
   hs-source-dirs:      src/
